==> code/3-Article_Recommendation_Content_Based_Approach.ipynb <==
import numpy as np
import pandas as pd
import dask.dataframe as dd
from scipy import stats
from matplotlib.pyplot import figure
from sklearn.metrics.pairwise import cosine_similarity
from scipy.sparse import csr_matrix
import warnings

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction import text

warnings.filterwarnings("ignore")
%matplotlib inline  

## Content Based Recomemndations

We try to related articles based on their category, garment group and description and try to build a sparse matrix mapping out their similarities. This is a non personalized recommendation but understanding similar products (articles in this case) is necessary to build a model.

path = "../files/"
articles_df = pd.read_csv(path + "articles.csv",
                          dtype={
                            "article_id": int,
                            "product_code": int,
                            "product_type_no": int,
                            "graphical_appearance_no": int,
                            "index_group_no": int,
                            "section_no": int,
                            "garment_group_no": int
                          })
articles_df.dropna(inplace = True)
articles_df.shape

articles_df['index_group_name'].value_counts()

## Similarities within Groups
Considering the Dataset is too large and similarities finding is very costly operation letz try to ascertain similarities 
within groups of the dataset using cosine similarity

Groups = articles_df['index_group_name'].value_counts().index
similarity_sparse_matrix = []
# Iterate groups from the H&M
for group in Groups:
    # Fetch the Articles Pertaining to the group
    df = articles_df[articles_df['index_group_name'] == group]
    # Concatenate the cotent for the particular product
    df['content_for_analyis'] = df['product_type_name'] + ' ' + articles_df['product_group_name'] + ' ' + articles_df['graphical_appearance_name'] + ' ' + articles_df['colour_group_name'] + ' ' + articles_df['department_name'] + ' ' + articles_df['section_name'] + ' ' + articles_df['garment_group_name']
    # Build the Dataframe for content based similarity index
    item_keywords_df = df[['article_id', 'content_for_analyis']]
    # TF IDF Vectorization
    vectorizer = TfidfVectorizer(min_df=3, max_df=0.7, stop_words=text.ENGLISH_STOP_WORDS)
    # Do a TD IDF Analysis
    X = vectorizer.fit_transform(item_keywords_df['content_for_analyis'])
    # Get the Vectorized_data
    vectorized_data = vectorizer.get_feature_names_out()
    # Print 
    print(group)
    print("-----")
    print("FeatureSet")
    print(vectorized_data)
    # Build the TF IDF Data Frame
    tfidf_df = pd.DataFrame(X.toarray(), columns=vectorizer.get_feature_names())
    tfidf_df.index = df['article_id']
    # Compute the Cosine Similarities
    cosine_similarity_array = cosine_similarity(tfidf_df)
    # Append it to the top level cosine similarilty array
    similarity_sparse_matrix.append(cosine_similarity_array)
    

len(similarity_sparse_matrix[0]), len(similarity_sparse_matrix[0][0])

# Finding the similarity in products - Processing Group "LadiesWear"
group = 'Ladieswear'
Ladieswear_df = pd.DataFrame(data=similarity_sparse_matrix[0], 
                             index=articles_df[articles_df['index_group_name'] == group]['article_id'].values,
                             columns=articles_df[articles_df['index_group_name'] == group]['article_id'].values)
Ladieswear_df

# Finding the similarity in products
grouped_dataframes = []
# Processing Group wise
Groups = articles_df['index_group_name'].value_counts().index
# Iterate groups from the H&M
for x, group in enumerate(Groups):
    df = pd.DataFrame(data=similarity_sparse_matrix[x], 
                                 index=articles_df[articles_df['index_group_name'] == group]['article_id'].values,
                                 columns=articles_df[articles_df['index_group_name'] == group]['article_id'].values)
    grouped_dataframes.append(df)

grouped_dataframes[2]

# Now we have made super easy find relationship between two articles
# For example the relationship of score of article "162074071" to "181448022" is as easy as
grouped_dataframes[2].loc[162074071][181448022]

arr1 = ['python3.6' , 'python2' ,'python3']
arr1.remove('python2')
arr1


similar_count = 10
tbl = []
for i, group in enumerate(Groups):
    for j, item in enumerate(grouped_dataframes[i].index):
        top_10_matched_items = grouped_dataframes[i][item].nlargest(n=10).index
        top10 = ",".join(str(v) for v in (list(top_10_matched_items[top_10_matched_items != item])))
        row = [item, top10]
        tbl.append(row)
    # Temporary Status Update    
    print('Processing Complete --- ', group)
        
article_similarity_df = pd.DataFrame(data=tbl, columns=['article_id', 'similar_articles'])
article_similarity_df


# Persist File to the file system
article_similarity_df.to_csv(path + 'article-content-similarity.csv')
