==> code/4-Customer_Recommendation_Content_Based_Approach.ipynb <==
import numpy as np
import pandas as pd
import dask.dataframe as dd
from scipy import stats
from matplotlib.pyplot import figure
from sklearn.metrics.pairwise import cosine_similarity
from scipy.sparse import csr_matrix
import warnings

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction import text
from sklearn.metrics.pairwise import cosine_similarity


from sklearn.cluster import KMeans
import pandas as pd
from sklearn.preprocessing import MinMaxScaler


warnings.filterwarnings("ignore")
%matplotlib inline  

## Customer Attribute Based Recommendation through clustering

We try to related articles based on their category, garment group and description and try to build a sparse matrix mapping out their similarities. This is a non personalized recommendation but understanding similar products (articles in this case) is necessary to build a model.

path = "../files/"


temp1_df = pd.read_csv(path+"article-content-similarity.csv")
temp2_df = pd.read_csv(path+"customers.csv")


customer_df = pd.read_csv(path+"customer_intelligence.csv")

feature_set = ['unique_items', 'unique_departments', 'preferred_sales_channel', 'preferred_week',
               'age', 'monetary_value']


km = KMeans(n_clusters=25)
y_predicted = km.fit_predict(customer_df[feature_set])
customer_df['cluster']=y_predicted
customer_df['cluster'].value_counts()

customer_df.index=customer_df['customer_id']
customer_df

def split_dataframe(df, chunk_size = 10000): 
    chunks = list()
    num_chunks = len(df) // chunk_size + 1
    for i in range(num_chunks):
        chunks.append(df[i*chunk_size:(i+1)*chunk_size])
    return chunks


def similarity_checker(df):
    tbl = []
    chunks = split_dataframe(df, chunk_size=50000)
    for chunk in chunks:
        cosine_similarity_array = cosine_similarity(chunk)
        similarity_df = pd.DataFrame(data=cosine_similarity_array, 
                                 index=chunk.index,
                                 columns=chunk.index)
        for j, item in enumerate(similarity_df.index):
            top_10_matched_items = similarity_df[item].nlargest(n=10).index
            top10 = ",".join(str(v) for v in (list(top_10_matched_items[top_10_matched_items != item])))
            row = [item, top10]
            tbl.append(row)
    return tbl


s_df = pd.DataFrame(similarity_checker(customer_df[customer_df['cluster'] == 24][feature_set]))
print(s_df.shape)

for i in range(24):
    t_df2 = pd.DataFrame(similarity_checker(customer_df[customer_df['cluster'] == i][feature_set]))
    print(i)
    print(t_df2.shape)
    s_df = s_df.append(t_df2, ignore_index=True)
    print(s_df.shape)


s_df.to_csv(path + 'customer-similarity.csv')

s_df
