==> code/5-finalconsolidation.ipynb <==
import numpy as np
import pandas as pd
import dask.dataframe as dd
from scipy import stats
from matplotlib.pyplot import figure
from sklearn.metrics.pairwise import cosine_similarity
from scipy.sparse import csr_matrix
import warnings
import random


warnings.filterwarnings("ignore")
%matplotlib inline  

path = "../files/"
#sample_customer_id1 = '00000dbacae5abe5e23885899a1fa44253a17956c6d1c3d25f88aa139fdfc657'

customer_df = pd.read_csv(path+"customers.csv", usecols=['customer_id'])
customer_df

# Read Merged CSV
customer_product_df = pd.read_csv(path + "customer_product_purchase.csv", usecols=['customer_id', 'article_id', 'count'])
customer_product_df.head(2)

missing_df = pd.merge(customer_df, customer_product_df, on="customer_id", how="outer")
missing_df

article_similarity_df = pd.read_csv(path+"article-content-similarity.csv", usecols=['article_id','similar_articles'])
article_similarity_df.head(2)

customers_similarity_df = pd.read_csv(path+"customer-similarity.csv", skiprows=1, names=['customer_id', 'similar_customers'])
customers_similarity_df.head(2)


def fn_similar_articles(article_id):
    lst =  article_similarity_df[article_similarity_df['article_id'] == article_id]['similar_articles'].tolist()
    s = ''
    if lst:
        s = lst[0]
    return s

fn_similar_articles(176209023)

def fn_similar_customers(customerid):
    lst =  customers_similarity_df[customers_similarity_df['customer_id'] == customerid]['similar_customers'].tolist()
    s = ''
    if lst:
        s = lst[0]
    return s


def fn_top_puchases(customerid):
    lst = customer_product_df[customer_product_df['customer_id'] == customerid].nlargest(3, 'count')['article_id'].tolist()
    return ",".join(str(x) for x in lst)


def predict_by_customers(customerid):
    articles_space_seperated = ''
    try:
        similar_customers =  customers_similarity_df[customers_similarity_df['customer_id'] == customerid]['similar_customers'].tolist()
        similar_customers_list = similar_customers[0].split(",")
        articles = []
        for customerid2 in similar_customers_list:
            t = fn_top_puchases(customerid2).split(",")
            articles.extend(t)
        articles_space_seperated = " ".join(str(x) for x in articles[0:12])
    except:
        print("Errored - ", customerid)
    return articles_space_seperated


def predict_by_purchase_history(customerid):
    articles_space_seperated = ''
    try:
        purchased_articles_list = fn_top_puchases(customerid).split(",")
        articles = []
        for item in purchased_articles_list:
            similar_articles =  fn_similar_articles(176209023)
            articles.extend(similar_articles.split(","))
        articles_space_seperated = " ".join(str(x) for x in articles[0:12])
    except:
        print("Errored - ", customerid)
    return articles_space_seperated
                  

k = 1500
arr = np.array_split(customer_df, k)

for i in range(k):
    chunk = pd.DataFrame(arr[i])
    chunk['predictions_by_customer'] = chunk['customer_id'].apply(lambda x: predict_by_customers(x))    
    chunk['predictions_by_purchase_history'] = chunk['customer_id'].apply(lambda x: predict_by_purchase_history(x))
    chunk.to_csv(path+"subfiles/intermediate" + str(i) + ".csv")

#customer_df['predictions_by_customer'] = customer_df['customer_id'].apply(lambda x: predict_by_customers(x))
#customer_df

#customer_df['predictions_by_purchase_history'] = customer_df['customer_id'].apply(lambda x: predict_by_purchase_history(x))
#customer_df


customer_df.to_csv(path+"intermediate1.csv")
